tender and authentication logic added

+++ src/main/java/coursework/backend/entity/Bid.java
@@ -39,7 +39,7 @@ public class Bid {
     private UUID tenderID;
 
     @Column(name = "cost", nullable = false)
-    private Double cost;
+    private Long cost;
 
     @Column(nullable = false)
     private String region;
